move board_list.dart to common_widgets directory
