+++ lib/view/common_widget/board_tile.dart
